+++ package.json
@@ -27,6 +27,7 @@
     "eslint": "^4.19.1"
   },
   "dependencies": {
-    "async": "^2.6.0"
+    "async": "^2.6.0",
+    "shelljs": "^0.8.1"
   }
 }
